travel_planner/requirements.txt
langchain-google-genai
tavily-python
python-dotenv